templated ofClamp version
use ofClamp instead of glm::clamp
to avoid inclusion of glm

+++ libs/openFrameworks/graphics/ofPolyline.inl
@@ -470,7 +470,7 @@ template<class T>
 ofPolyline_<T> ofPolyline_<T>::getSmoothed(int smoothingSize, float smoothingShape) const {
 	int n = size();
 	smoothingSize = ofClamp(smoothingSize, 0, n);
-	smoothingShape = ofClamp(smoothingShape, 0, 1);
+	smoothingShape = ofClamp(smoothingShape, 0.f, 1.f);
 	
 	// precompute weights and normalization
 	std::vector<float> weights;
@@ -891,17 +891,17 @@ float ofPolyline_<T>::getIndexAtLength(float length) const {
     updateCache();
     
     float totalLength = getPerimeter();
-    length = ofClamp(length, 0, totalLength);
+    length = ofClamp(length, 0.f, totalLength);
     
     int lastPointIndex = isClosed() ? points.size() : points.size()-1;
     
-    int i1 = ofClamp(floor(length / totalLength * lastPointIndex), 0, lengths.size()-2);   // start approximation here
+    int i1 = ofClamp(int(floor(length / totalLength * lastPointIndex)), 0, int(lengths.size()-2));   // start approximation here
     int leftLimit = 0;
     int rightLimit = lastPointIndex;
     
     float distAt1, distAt2;
     for(int iterations = 0; iterations < 32; iterations ++) {	// limit iterations
-        i1 = ofClamp(i1, 0, lengths.size()-1);
+        i1 = ofClamp(i1, 0, int(lengths.size()-1));
         distAt1 = lengths[i1];
         if(distAt1 <= length) {         // if Length at i1 is less than desired Length (this is good)
             distAt2 = lengths[i1+1];
